requirements.txt
Flask==3.0.0
google-generativeai
PyPDF2
langchain-text-splitters
python-dotenv
beautifulsoup4
pdfplumber
mistralai
nameparser
pandas
requests